requirements.txt
websocket-client
redis
numpy
pandas
requests
python-dotenv
